+++ examples/nest/src/bull-horizon/bull-horizon.service.ts
@@ -1,8 +1,8 @@
-import { BullMonitorExpress } from '@bull-monitor/express';
+import { BullMonitorExpress } from '@bull-horizon/express';
 import { InjectQueue } from '@nestjs/bull';
 import { Injectable } from '@nestjs/common';
 import { Queue } from 'bull';
-import { BullAdapter } from '@bull-monitor/root/dist/bull-adapter';
+import { BullAdapter } from '@bull-horizon/root/dist/bull-adapter';
 
 @Injectable()
 export class BullMonitorService extends BullMonitorExpress {